==> step1_building_patient_df.ipynb <==
# Machine Learning - Predicting Treatment Abandonment with scikit learn
By **Daniel Palacio** (github.com/palaciodaniel) - 2020

## STEP ONE - Building the DataFrame

In order to build the table, we will need to have the Faker package installed. For obvious privacy reasons, the use of real patient data has been ruled out, so we will build a fictional DataFrame from scratch. In case you do not have the Faker package installed, please remove the "#" sign on the following cell and Run it.

**DO NOT RUN THE REST OF THE CELLS, OTHERWISE THEIR OUTPUT WILL BE LOST.**

# pip install Faker

1.  We import Faker and "random" libraries. We will use Faker to create two columns: "Name" and "Gender" ([M]ale and [F]emale). As for "random", we will use it to generate a column with ages from 18 to 85.

    All the information will be stored in a list of lists (with every nested list representing a single observation). We will generate 100 observations, so the list should have that length.

    Notice that for both Faker and "random" we will be using seeds, so that we obtain the same results every time.

from faker import Faker
import random

fake = Faker()
Faker.seed(24)

names_list = []

for i in range(100):
    random.seed(i)
    profile = fake.simple_profile()
    names_list.extend([[profile["name"], profile["sex"], random.randint(18, 85)]])

print("List length:", len(names_list), "\n")
print(names_list)

2.  We will now generate columns for different fictional features. The use of a real test was dismissed to avoid potential copyright infringements, but still, this whole demonstration shows how it is perfectly feasible to implement a Machine Learning approach on a situation like this.

    Obviously, every observation feature will be filled at "random", from a list of three possible choices: "Low", "Medium" and "High".
    
    A brief description of some of the non-obvious (and obviously fictional) features:
    
    * **Neuroticism:** Indicates how much the adaptation to everyday's life demands was compromised.
    * **Resourcefulness:** This one not only includes intelligence, but also any other resource that can be used to fulfill the treatment's objectives (i.e.: salary, social skills, etc.).
    * **Social Expectation:** Measures the perceived level of "pressure" the patient has from the most immediate social circles to overcome the problem/mental disorder.
    * **Introspection:** The ability to understand and assess all the changes the treatment produces on the patient's well-being and general quality of life.
    * **Victimhood:** Unlike the previous features, this one is either "True" or "False". If "True", the patient tends to believe life will always be inherently "harsh", while people on the opposite side tend to be more "combative" and understand that change is possible.

neuroticism = []
motivation = []
resourcefulness = []
social_expectation = []
introspection = []
discipline = []
victimhood = []

for i in range(100):
    random.seed(i)
    neuroticism.append(random.choice(["Low", "Medium", "High"]))
    motivation.append(random.choice(["Low", "Medium", "High"]))
    resourcefulness.append(random.choice(["Low", "Medium", "High"]))
    social_expectation.append(random.choice(["Low", "Medium", "High"]))
    introspection.append(random.choice(["Low", "Medium", "High"]))
    discipline.append(random.choice(["Low", "Medium", "High"]))
    victimhood.append(random.choice([True, False]))

3. Now we import the pandas module and transform the nested lists from Step 1 into a proper DataFrame. 

##### NOTE: The column 'Genre' should be 'Gender' or 'Sex'. This is a typing mistake. However, it was left like that to avoid having to re-execute Step 7, but it will be fixed for the second phase.

import pandas as pd

patient_df = pd.DataFrame(names_list, columns=["Name", "Genre", "Age"])
print(patient_df.head())

4. We also integrate the features into a DataFrame as well.

patient_df2 = pd.DataFrame(list(zip(neuroticism, motivation, resourcefulness,
                                    social_expectation, introspection, discipline, victimhood)),
                          columns = ["Neuroticism", "Motivation", "Resourcefulness", "Social Expectation",
                                     "Introspection", "Discipline", "Victimhood"])

print("DataFrame's shape:", patient_df2.shape, "\n")
print(patient_df2.head())

5. We combine the DataFrames from both previous steps into a single one.

patient_final_df = pd.concat([patient_df, patient_df2], axis = 1)

print("Final DataFrame's shape:", patient_final_df.shape, "\n")
print(patient_final_df.head())

6.  We are ready to create the target column now. The following code prints every single observation and allows to input a "realistic" treatment's outcome based on all features. Did the patient finish the treatment?

    After examining all the 100 observations, the selected outcomes will be appended to a list.

finished_treatment = []

for i in range(len(patient_final_df)):
    print(patient_final_df.loc[i, :])
    your_input = input("Finished treatment? (TRUE/FALSE): ")
    finished_treatment.append(your_input)

print(finished_treatment)   

7. We append the target column to the previous combined DataFrame. It is officially finished now.

patient_final_df["Finished"] = finished_treatment
print(patient_final_df.head())

8. We save the DataFrame to a CSV file, ready to be cleaned and prepared on Step 2...

patient_final_df.to_csv("df_patients.csv", index = True)
