==> step2_cleaning_the_dataset.ipynb <==
# Machine Learning - Predicting Treatment Abandonment with scikit learn¶
By **Daniel Palacio** (github.com/palaciodaniel) - 2020

## STEP TWO - Adapting the DataFrame

Having successfully created a functional DataFrame about patients's personality attributes, and whether they managed to finish their treatments or not, now it is time to clean the DataFrame so that it becomes suitable for Machine Learning.

Basically, the challenge on this step is to transform all column values to numeric ones, otherwise we will not be able to fit them to our Machine Learning model. But as we will see now, there is only one column ("Age") that currently satisfies that requirement...

# Loading DataFrame into variable 'patients_df'

import pandas as pd

patients_df = pd.read_csv("df_patients.csv", header = 0, index_col = 0)

patients_df.head()

# Check data types on every column from 'patients_df'

patients_df.info()

# Force columns with boolean values into integers (therefore 0 = False and 1 = True)

patients_df[["Victimhood", "Finished"]] = patients_df[["Victimhood", "Finished"]].astype("int")
patients_df.head()

# Dropping column 'Name' (this one isn't necessary for the model)

patients_df = patients_df.drop("Name", axis = 1)
patients_df.head()

# Transforming categorical columns into dummies

df_dummies = pd.get_dummies(patients_df[["Genre", "Neuroticism", "Motivation", "Resourcefulness", \
                   "Social Expectation", "Introspection", "Discipline"]])

print(df_dummies)

# Dropping original categorical columns from 'patients_df'

patients_df = patients_df.drop(["Genre", "Neuroticism", "Motivation", "Resourcefulness", \
                   "Social Expectation", "Introspection", "Discipline"], axis = 1)
patients_df.head()

# Concatenating two remaining columns of 'patients_df' with 'df_dummies'.

prepared_df = pd.concat([patients_df, df_dummies], axis = 1)
print(prepared_df.head())
print("DataFrame's shape: ", prepared_df.shape)

# Removing redundant column "Genre_F"

prepared_df = prepared_df.drop("Genre_F", axis = 1)
prepared_df.head()

# Reordering dummy columns so that new order is 'Low', 'Medium' and then 'High'.
# Also, target column "Finished" will be moved to the very end.

prepared_df = prepared_df[["Age", "Genre_M", "Victimhood",
                           "Discipline_Low", "Discipline_Medium", "Discipline_High", 
                           "Introspection_Low", "Introspection_Medium", "Introspection_High",
                           "Motivation_Low", "Motivation_Medium", "Motivation_High",
                           "Neuroticism_Low", "Neuroticism_Medium", "Neuroticism_High",
                           "Resourcefulness_Low", "Resourcefulness_Medium", "Resourcefulness_High",
                           "Social Expectation_Low", "Social Expectation_Medium", "Social Expectation_High",
                           "Finished"]]
prepared_df.info()

# Renaming 'Genre_M' to 'Sex_Male'
prepared_df.columns = [col.replace('Genre_M', 'Sex_Male') for col in prepared_df.columns]
print(prepared_df["Sex_Male"][:5])

# Saving DataFrame 'prepared_df' to a CSV file

prepared_df.to_csv("df_prepared.csv", index = True)
