==> step4_improving_ml_with_gridsearch.ipynb <==
# Machine Learning - Predicting Treatment Abandonment with scikit learn
By **Daniel Palacio** (github.com/palaciodaniel) - 2020

## STEP FOUR - Improving model with GridSearch
Considering the dataset involves a sensitive subject (mental health) it becomes mandatory to look for the best model, the one with the best score. And since we are dealing with a very small dataset, then we can use GridSearch, so that we can determine which hyperparameters are the most suitable for the case.


# Importing required libraries

from sklearn.linear_model import LogisticRegression

from sklearn.preprocessing import StandardScaler

from sklearn.model_selection import GridSearchCV # This is the new one we are importing.
from sklearn.model_selection import train_test_split

from sklearn.metrics import confusion_matrix
from sklearn.metrics import accuracy_score,recall_score,precision_score
import pandas as pd

# Loading prepared DataFrame (only numeric values)

df = pd.read_csv("df_prepared.csv", header = 0, index_col = 0)

df.info()

### Final preparations

# Feature columns

X = df.iloc[:,:-1].to_numpy()
print(X[:5])

# Target column

y = df.iloc[:, -1].to_numpy()
print(y[:5])

# Dividing between training and test subsets

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.2, shuffle = True, random_state = 24)

print("X_train:", X_train.shape, type(y_test))
print("X_test:", X_test.shape, type(y_test))
print("y_train:", y_train.shape, type(y_test))
print("y_test:", y_test.shape, type(y_test))

# Scaling the data

scaler = StandardScaler()
X_train_scaled = scaler.fit_transform(X_train)
X_test_scaled = scaler.transform(X_test)

Up to this point there was not a big difference with Step 3, but now we are finally applying GridSearch...

### Model application

**IMPORTANT NOTE:** Logistic Regression has more parameters than the ones selected here; however, given that GridSearch looks for all possible combinations, in this case some of them weren't compatible, therefore it was pointless to spend processing time only to output a long list of warnings, and that is the main reason why some of them were ruled out.

Also, scikit learn's __[documentation for Logistic Regression](https://scikit-learn.org/stable/modules/generated/sklearn.linear_model.LogisticRegression.html)__ specifically recommends the solver 'liblinear' for small datasets.

# Instantiating and fitting a Logistic Regression model with GridSearch

log_reg = LogisticRegression(random_state = 24)

params = {
        "penalty": ["l1", "l2"], 
        "C": [0.01, 0.1, 0.25, 0.5, 0.75, 1.0, 1.25, 1.5, 1.75, 2.0, 2.5, 5, 10], 
        "solver": ["liblinear"]
         }

grid_search = GridSearchCV(log_reg, param_grid = params)

grid_search.fit(X_train_scaled, y_train)

# Making predictions, now with Grid Search

y_pred = grid_search.predict(X_test_scaled)
print("y_pred:", y_pred.shape, y_pred)
print("y_test:", y_test.shape, y_test)

### Scores

# Confusion Matrix
conf_matrix = confusion_matrix(y_test, y_pred)
print(conf_matrix)

tn, fp, fn, tp = confusion_matrix(y_test, y_pred).ravel()
print("True Positives:", tp)
print("True Negatives:", tn)
print("False Positives:", fp)
print("False Negatives:", fn)

# The best parameters

print("GridSearch revealed that the best parameters were the following:\n", grid_search.best_params_)

# New accuracy, precision and recall scores

print("\nThe new (and improved) scores are the following:")
print("Accuracy:", round(accuracy_score(y_test,y_pred), 2))
print("Precision", round(precision_score(y_test,y_pred), 2))
print("Recall:", round(recall_score(y_test,y_pred), 2))

### Conclusion
Using GridSearch definitely improved the scores; it is far from being a robust model, of course, and even more considering it is about a delicate subject, as it was already expressed before. However, if we think that we used a dataset containing only 100 observations, it is already a promising start. 

Obviously, if we were going to continue improving the model, we would need to add much more observations. Thousands of them, to be precise.
